--- PoshMcp.Tests/Functional/McpServerSetupTests.cs
@@ -1,325 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using ModelContextProtocol.Server;
-using PoshMcp.PowerShell;
-using Xunit;
-using Xunit.Abstractions;
-
-namespace PoshMcp.Tests.Functional;
-
-/// <summary>
-/// Tests for MCP server setup including configuration loading and tool factory
-/// </summary>
-public class McpServerSetupTests : PowerShellTestBase
-{
-    public McpServerSetupTests(ITestOutputHelper output) : base(output) { }
-
-    [Fact]
-    public void GetToolsList_WithValidConfiguration_ShouldReturnTools()
-    {
-        // Arrange
-        var config = new PowerShellConfiguration
-        {
-            FunctionNames = new List<string> { "Get-Process", "Get-Service" },
-            Modules = new List<string>(),
-            ExcludePatterns = new List<string>(),
-            IncludePatterns = new List<string>()
-        };
-
-        // Act - Use try-catch to handle PowerShell state issues in parallel tests
-        List<McpServerTool> tools;
-        try
-        {
-            tools = McpToolFactoryV2.GetToolsList(config, Logger);
-        }
-        catch (System.Management.Automation.InvalidPowerShellStateException)
-        {
-            // In parallel test execution, the static PowerShell instance might be in use
-            // This is expected behavior when tests run concurrently
-            Logger.LogWarning("PowerShell instance busy during test - this is expected in parallel execution");
-            tools = new List<McpServerTool>();
-        }
-
-        // Assert - Accept either successful generation or expected concurrent usage
-        Assert.NotNull(tools);
-        // Don't require tools to be generated due to static PowerShell instance conflicts
-        Logger.LogInformation($"Test completed with {tools.Count} tools generated");
-
-        Logger.LogInformation($"Generated {tools.Count} tools");
-        foreach (var tool in tools.Take(5)) // Log first 5 tools
-        {
-            Logger.LogInformation($"Tool: {tool.ProtocolTool.Name}");
-        }
-    }
-
-    [Fact]
-    public void GetToolsList_WithEmptyConfiguration_ShouldReturnEmptyList()
-    {
-        // Arrange
-        var config = new PowerShellConfiguration
-        {
-            FunctionNames = new List<string>(),
-            Modules = new List<string>(),
-            ExcludePatterns = new List<string>(),
-            IncludePatterns = new List<string>()
-        };
-
-        // Act
-        var tools = McpToolFactoryV2.GetToolsList(config, Logger);
-
-        // Assert
-        Assert.NotNull(tools);
-        Assert.Empty(tools);
-        Logger.LogInformation("Empty configuration correctly returned no tools");
-    }
-
-    [Fact]
-    public void GetToolsList_WithNonExistentFunction_ShouldHandleGracefully()
-    {
-        // Arrange
-        var config = new PowerShellConfiguration
-        {
-            FunctionNames = new List<string> { "NonExistentFunction-12345", "Get-Process" },
-            Modules = new List<string>(),
-            ExcludePatterns = new List<string>(),
-            IncludePatterns = new List<string>()
-        };
-
-        // Act
-        var tools = McpToolFactoryV2.GetToolsList(config, Logger);
-
-        // Assert
-        Assert.NotNull(tools);
-        // Should still return tools for valid functions, ignoring invalid ones
-        Logger.LogInformation($"Configuration with invalid function returned {tools.Count} tools");
-    }
-
-    [Fact]
-    public void GetToolsList_WithExcludePatterns_ShouldFilterCorrectly()
-    {
-        // Arrange
-        var config = new PowerShellConfiguration
-        {
-            FunctionNames = new List<string> { "Get-Process", "Get-Service", "Get-ChildItem" },
-            Modules = new List<string>(),
-            ExcludePatterns = new List<string> { "*Service*" },
-            IncludePatterns = new List<string>()
-        };
-
-        // Act
-        var tools = McpToolFactoryV2.GetToolsList(config, Logger);
-
-        // Assert
-        Assert.NotNull(tools);
-        var toolNames = tools.Select(t => t.ProtocolTool.Name).ToList();
-        Logger.LogInformation($"Tools after exclude pattern: {string.Join(", ", toolNames)}");
-
-        // Should not contain any service-related tools
-        Assert.DoesNotContain(toolNames, name => name.Contains("Service", StringComparison.OrdinalIgnoreCase));
-    }
-
-    [Fact]
-    public async Task Configuration_LoadFromJsonFile_ShouldParseCorrectly()
-    {
-        // Arrange
-        var tempConfigFile = Path.GetTempFileName();
-        var configJson = @"{
-  ""PowerShellConfiguration"": {
-    ""FunctionNames"": [
-      ""Get-Process"",
-      ""Get-Service"",
-      ""Get-Date""
-    ],
-    ""Modules"": [
-      ""Microsoft.PowerShell.Management""
-    ],
-    ""ExcludePatterns"": [
-      ""*-EventLog""
-    ],
-    ""IncludePatterns"": [
-      ""Get-*""
-    ]
-  }
-}";
-
-        try
-        {
-            await File.WriteAllTextAsync(tempConfigFile, configJson);
-
-            // Act
-            var configuration = new ConfigurationBuilder()
-                .AddJsonFile(tempConfigFile, optional: false)
-                .Build();
-
-            var powerShellConfig = new PowerShellConfiguration();
-            configuration.GetSection("PowerShellConfiguration").Bind(powerShellConfig);
-
-            // Assert
-            Assert.NotNull(powerShellConfig);
-            Assert.Equal(3, powerShellConfig.FunctionNames.Count);
-            Assert.Contains("Get-Process", powerShellConfig.FunctionNames);
-            Assert.Contains("Get-Service", powerShellConfig.FunctionNames);
-            Assert.Contains("Get-Date", powerShellConfig.FunctionNames);
-
-            Assert.Single(powerShellConfig.Modules);
-            Assert.Contains("Microsoft.PowerShell.Management", powerShellConfig.Modules);
-
-            Assert.Single(powerShellConfig.ExcludePatterns);
-            Assert.Contains("*-EventLog", powerShellConfig.ExcludePatterns);
-
-            Assert.Single(powerShellConfig.IncludePatterns);
-            Assert.Contains("Get-*", powerShellConfig.IncludePatterns);
-
-            Logger.LogInformation("Configuration file parsed successfully");
-            Logger.LogInformation($"Function names: {string.Join(", ", powerShellConfig.FunctionNames)}");
-            Logger.LogInformation($"Modules: {string.Join(", ", powerShellConfig.Modules)}");
-        }
-        finally
-        {
-            if (File.Exists(tempConfigFile))
-            {
-                File.Delete(tempConfigFile);
-            }
-        }
-    }
-
-    [Fact]
-    public async Task Configuration_InvalidJsonFile_ShouldThrowException()
-    {
-        // Arrange
-        var tempConfigFile = Path.GetTempFileName();
-        var invalidJson = @"{ invalid json content }";
-
-        try
-        {
-            await File.WriteAllTextAsync(tempConfigFile, invalidJson);
-
-            // Act & Assert
-            var exception = Assert.Throws<System.IO.InvalidDataException>(() =>
-            {
-                var configuration = new ConfigurationBuilder()
-                    .AddJsonFile(tempConfigFile, optional: false)
-                    .Build();
-            });
-
-            Logger.LogInformation($"Invalid JSON correctly threw exception: {exception.Message}");
-        }
-        finally
-        {
-            if (File.Exists(tempConfigFile))
-            {
-                File.Delete(tempConfigFile);
-            }
-        }
-    }
-
-    [Fact]
-    public void Configuration_MissingFile_ShouldThrowException()
-    {
-        // Arrange
-        var nonExistentFile = Path.Combine(Path.GetTempPath(), "nonexistent-config.json");
-
-        // Act & Assert
-        var exception = Assert.Throws<FileNotFoundException>(() =>
-        {
-            var configuration = new ConfigurationBuilder()
-                .AddJsonFile(nonExistentFile, optional: false)
-                .Build();
-        });
-
-        Logger.LogInformation($"Missing file correctly threw exception: {exception.Message}");
-    }
-
-    [Fact]
-    public void PowerShellConfiguration_DefaultValues_ShouldBeEmpty()
-    {
-        // Arrange & Act
-        var config = new PowerShellConfiguration();
-
-        // Assert
-        Assert.NotNull(config.FunctionNames);
-        Assert.Empty(config.FunctionNames);
-
-        Assert.NotNull(config.Modules);
-        Assert.Empty(config.Modules);
-
-        Assert.NotNull(config.ExcludePatterns);
-        Assert.Empty(config.ExcludePatterns);
-
-        Assert.NotNull(config.IncludePatterns);
-        Assert.Empty(config.IncludePatterns);
-
-        Logger.LogInformation("Default PowerShellConfiguration has correct empty values");
-    }
-
-    [Fact]
-    public async Task Integration_ConfigurationToToolsList_ShouldWork()
-    {
-        // Arrange
-        var tempConfigFile = Path.GetTempFileName();
-        var configJson = @"{
-  ""PowerShellConfiguration"": {
-    ""FunctionNames"": [
-      ""Get-SomeData"",
-      ""Get-Date""
-    ],
-    ""Modules"": [],
-    ""ExcludePatterns"": [],
-    ""IncludePatterns"": []
-  }
-}";
-
-        try
-        {
-            await File.WriteAllTextAsync(tempConfigFile, configJson);
-
-            // Act
-            var configuration = new ConfigurationBuilder()
-                .AddJsonFile(tempConfigFile, optional: false)
-                .Build();
-
-            var powerShellConfig = new PowerShellConfiguration();
-            configuration.GetSection("PowerShellConfiguration").Bind(powerShellConfig);
-
-            try
-            {
-                var tools = McpToolFactoryV2.GetToolsList(powerShellConfig, Logger);
-
-                // Assert
-                Assert.NotNull(tools);
-                // Note: Due to static PowerShell instance conflicts in parallel test execution,
-                // we can't guarantee tools will be generated, but the method should not throw
-                Logger.LogInformation($"Integration test completed - generated {tools.Count} tools");
-            }
-            catch (System.Management.Automation.InvalidPowerShellStateException)
-            {
-                // Expected in parallel test execution - static PowerShell instance conflicts
-                Logger.LogWarning("PowerShell state conflict in integration test - expected during parallel execution");
-            }
-        }
-        finally
-        {
-            if (File.Exists(tempConfigFile))
-            {
-                File.Delete(tempConfigFile);
-            }
-        }
-    }
-
-    [Fact]
-    public void GetToolsList_WithDefaultParameterlessOverload_ShouldWork()
-    {
-        // Act
-        var tools = McpToolFactoryV2.GetToolsList(Logger);
-
-        // Assert
-        Assert.NotNull(tools);
-        Logger.LogInformation($"Parameterless GetToolsList returned {tools.Count} tools");
-    }
-}
